Pequenas correções

+++ src/main/java/ufg/pw/projeto_restaurante/model/utils/endereco/dao/LogradouroDao.java
@@ -3,7 +3,7 @@ package ufg.pw.projeto_restaurante.model.utils.endereco.dao;
 import ufg.pw.projeto_restaurante.model.utils.endereco.Logradouro;
 import ufg.pw.projeto_restaurante.model.utils.persistencia.GenericDao;
 
-public class LogradouroDao extends GenericDao<Logradouro, Integer> {
+public class LogradouroDao extends GenericDao<Logradouro, Long> {
 	public LogradouroDao() {
 		super(Logradouro.class);
 	}
